+++ NEWS.md
@@ -3,3 +3,13 @@
 ## V1.0.0
 
 * Initial release to CRAN
+
+## V1.0.1
+
+* Fixed bug which caused an error if a deforestation criterion only finds one cluster before the deforestation stage 
+
+* Fixed bug with memoisation and allowed a parameter to be specified which gives the user control of when to use RIBIS
+
+* Changed name of balanced deforestation criterion to "diverse"
+
+* Allowed UNCOVER to predict a single observation